fixup

+++ bootstrap/tests/test.camera.js
@@ -25,6 +25,8 @@ var tj = new TJBot(['camera'], {}, {});
 
 tj._captureImage('picture.jpg').then(function(data) {
     assert(fs.existsSync('picture.jpg'), "expected picture.jpg to have been created");
-    fs.unlink('picture.jpg');
+    if (fs.existsSync('picture.jpg')) {
+        fs.unlink('picture.jpg');
+    }
     assert(!fs.existsSync('picture.jpg'), "expected to have deleted picture.jpg");
 });
